Revert "Merge pull request #7 from dreamyPatisiel/DP-36"
This reverts commit 65ebca9e5af12b7a8f5bf4de35884b58fdb2a04b, reversing
changes made to 21e44f317312646a49e2a31e4d97ccf98124f9a7.
