+++ package.json
@@ -14,7 +14,6 @@
     "@tanstack/react-query": "^5.14.2",
     "@tanstack/react-query-devtools": "^5.17.21",
     "axios": "^1.6.5",
-    "framer-motion": "^11.0.3",
     "next": "14.0.4",
     "next-auth": "^4.24.5",
     "next-themes": "^0.2.1",
