+++ pages/index.tsx
@@ -1,9 +1,6 @@
-import Modal from '@/components/modal';
-import { useModalStore } from '@/store/modalStore';
-
 export default function Index() {
   return (
-    <div className='text-white'>
+    <div className=''>
       <h1>DEVDEVDEV</h1>
       <h2 className='font-bold '>ddd</h2>
       <p className='text-st1 font-light leading-st'>hello</p>
